% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albersusa-package.R
\docType{package}
\name{albersusa}
\alias{albersusa}
\alias{albersusa-package}
\title{Tools, Shapefiles & Data to Work with an 'AlbersUSA' Composite Projection}
\description{
Tools, Shapefiles & Data to Work with an 'AlbersUSA' Composite Projection
}
\author{
Bob Rudis (@hrbrmstr)
}

