% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_elided.R
\name{points_elided}
\alias{points_elided}
\title{Shift points around Alaska and Hawaii to the elided area}
\usage{
points_elided(sp)
}
\arguments{
\item{sp}{An object of SpatialPoints class}
}
\value{
An elided version of the original SpatialPoints class
}
\description{
This function will take a SpatialPoints object and shift the points around
 Alaska and Hawaii to the elided area from this package.
}
