% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points-elided-sf.R
\name{points_elided_sf}
\alias{points_elided_sf}
\title{Shift points around Alaska and Hawaii to the elided area}
\usage{
points_elided_sf(sfin)
}
\arguments{
\item{{sfin}}{An object of SpatialPoints class or a data frame with x (`lon`) and y (`lat`)}
}
\value{
An elided version of the original {sf} class
}
\description{
This function will take an {sf} object or a data frame of coordinates
and shift the points around Alaska and Hawaii to the elided area from this package,
leaving the other points intact.
}
