% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.r
\docType{data}
\name{us_aeqd_proj}
\alias{us_aeqd_proj}
\title{Oblique azimuthal equidistant convenience projection}
\format{An object of class \code{character} of length 1.}
\usage{
us_aeqd_proj
}
\description{
Distance correct along any line passing through the center of the map (i.e., great circle)
}
\details{
The following PROJ.4 string is used: \code{+proj=aeqd +lat_0=40.08355112574181 +lon_0=-95.44921875}

In a ggplot2+ggalt context, you can just use variable name as the value for the \code{proj}
parameter to \code{ggalt::coord_proj()}. i.e. \code{coord_proj(us_aeqd_proj)}.
}
\examples{
us <- usa_composite()
us <- sp::spTransform(us, sp::CRS(us_aeqd_proj))
us <- usa_composite(proj="aeqd")
counties <- counties_composite("aeqd")
}
\seealso{
Other convenience projections: 
\code{\link{us_eqdc_proj}},
\code{\link{us_laea_proj}},
\code{\link{us_lcc_proj}}
}
\concept{convenience projections}
\keyword{datasets}
