% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.r
\docType{data}
\name{us_lcc_proj}
\alias{us_lcc_proj}
\title{Lambert conformal conic convenience projection}
\format{An object of class \code{character} of length 1.}
\usage{
us_lcc_proj
}
\description{
The following PROJ.4 string is used: \code{+proj=lcc +lat_1=27.11637320883929 +lat_2=53.050729042644335 +lon_0=-95.44921875}
}
\details{
In a ggplot2+ggalt context, you can just use variable name as the value for the \code{proj}
parameter to \code{ggalt::coord_proj()}. i.e. \code{coord_proj(us_lcc_proj)}.
}
\examples{
us <- usa_composite()
us <- sp::spTransform(us, sp::CRS(us_lcc_proj))
us <- usa_composite(proj="lcc")
counties <- counties_composite("lcc")
}
\seealso{
Other convenience projections: 
\code{\link{us_aeqd_proj}},
\code{\link{us_eqdc_proj}},
\code{\link{us_laea_proj}}
}
\concept{convenience projections}
\keyword{datasets}
