% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albersusa.R
\name{usa_composite}
\alias{usa_composite}
\title{Retreive a U.S. composite map, optionally with a projection}
\usage{
usa_composite(proj = c("longlat", "laea", "lcc", "eqdc", "aeqd"))
}
\arguments{
\item{proj}{if anything but \code{longlat} is specified, the shapefile will
be pre-projected before being returned}
}
\description{
The \code{proj} parameter is intende to simplify usage for those using base
plotting. If anything but \code{longlat} is specified the composite map
will be pre-projected to those
}
\examples{
plot(usa_composite())

plot(usa_composite("laea"))
}

